Define paths using PATH_MAX
Use the defined value PATH_MAX instead of an harcoded value
because it can differ from one operating system to another.

I know that it's use is ambiguous at this point because it might not
be enforced on some systems. However I believe that it shouldn't be
too ambigious for the needs of this utility.