renamer_file.c: Change return type of the string manipulation functions
It is kinda weird to use 2 differents parameters in order to modify
a string. (One source which is being modified into the destination)

For this reason modify the return type of remove_headers() and
remove_whitespaces().

Since it might be possible that the said string is cleaned up from
the stack after the value has been returned, save it in the heap
first using strdup().