Adjust hsrenamer for future fansubs
Whilst renaming HorribleSubs is good, it might be a little better
to use hsrenamer to rename more than one fansub at the time no?

Separate the renaming process to a new file renamer_fansubs in which
each related fansubs will have their own remove_header() and
remove_whitespaces() functions.

It might not be the most flexible code to date since duplication
is probably to be expected with the increasing number of fansubs.