Adapt hsrenamer for most releases
Our last attempt of trying to get given fansubs working would cause
a problem since there are a lot of them. Copying code and
implementing each of the given functions would be tedious and hard
to debug in the long run.

In this case, since most of the fansubs have either brackets ("[]")
and/or parentheses, ("()") only remove them as well as their given
sections.

Use renamer_parser that will remove all of these and give back
a "cleaned up" filename.