See farther, AMS!
As this initially touched only HorribleSubs releases, I thought
it would be a bit more versatile if this utility would touch
most of the anime releases available.

That's why we'll rename hsrenamer to amsrenamer so it doesn't
only touch HorribleSubs but also some other fansubs.

I don't personally think that a lot of people will use it as there
are much more powerful tools out there. That said it still was
a nice little project to work on.